+++ tests/test_typeguard_py36.py
@@ -2,8 +2,22 @@ import warnings
 from typing import AsyncGenerator, AsyncIterable, AsyncIterator
 
 import pytest
+from typing_extensions import Protocol, runtime_checkable
 
 from typeguard import TypeChecker, typechecked
+
+try:
+    from typing_extensions import TypedDict
+except ImportError:
+    from typing import TypedDict
+
+
+@runtime_checkable
+class RuntimeProtocol(Protocol):
+    member: int
+
+    def meth(self) -> None:
+        ...
 
 
 class TestTypeChecked:
@@ -78,6 +92,20 @@ class TestTypeChecked:
 
         foo()
 
+    def test_typeddict_inherited(self):
+        class ParentDict(TypedDict):
+            x: int
+
+        class ChildDict(ParentDict, total=False):
+            y: int
+
+        @typechecked
+        def foo(arg: ChildDict):
+            pass
+
+        foo({'x': 1})
+        pytest.raises(TypeError, foo, {'y': 1})
+
 
 async def asyncgenfunc() -> AsyncGenerator[int, None]:
     yield 1
@@ -106,3 +134,17 @@ class TestTypeChecker:
             func()
 
         assert len(record) == 0
+
+
+def test_protocol_non_method_members():
+    @typechecked
+    def foo(a: RuntimeProtocol):
+        pass
+
+    class Foo:
+        member = 1
+
+        def meth(self) -> None:
+            pass
+
+    foo(Foo())